+++ geoextent/lib/helpfunctions.py
@@ -1,36 +1,50 @@
-import sys, os, platform, datetime, math, random
-import zipfile, re
-from os.path import basename
-import pandas as pd
+import csv
+import datetime
+import itertools
+import logging
+import os
+import random
 import re
-from pandas.core.tools.datetimes import _guess_datetime_format_for_array as time_format
+import zipfile
 import numpy as np
+import pandas as pd
 from osgeo import ogr
 from osgeo import osr
-import logging
-from pyproj import Proj, transform
-import csv
+from pandas.core.tools.datetimes import _guess_datetime_format_for_array as time_format
+from pathlib import Path
 
 output_time_format = '%Y-%m-%d'
 PREFERRED_SAMPLE_SIZE = 30
 WGS84_EPSG_ID = 4326
 logger = logging.getLogger("geoextent")
 
+https_regexp = re.compile('https://(.*)')
 
-def getAllRowElements(rowname, elements, exp_data=None):
-    '''
+# doi_regexp, is_doi, and normalize_doi are from idutils (https://github.com/inveniosoftware/idutils)
+# Copyright (C) 2015-2018 CERN.
+# Copyright (C) 2018 Alan Rubin.
+# Licensed under BSD-3-Clause license
+doi_regexp = re.compile(
+    r"(doi:\s*|(?:https?://)?(?:dx\.)?doi\.org/)?(10\.\d+(.\d+)*/.+)$", flags=re.I)
+
+zenodo_regexp = re.compile(
+    r"(https://zenodo.org/record/)?(.\d*)$", flags=re.I
+)
+
+
+def getAllRowElements(row_name, elements, exp_data=None):
+    """
     Function purpose: help-function to get all row elements for a specific string \n
-    Input: rowname, elements, exp_format \n
+    Input: row name, elements, exp_format \n
     Output: array values
-    '''
-
+    """
+    values = []
     for idx, val in enumerate(elements[0]):
-        if rowname in val:
+        if row_name in val:
             indexOf = idx
-            values = []
             for x in elements:
                 try:
-                    if x[indexOf] != rowname:
+                    if x[indexOf] != row_name:
                         values.append(x[indexOf].replace(" ", ""))
                 except IndexError as e:
                     logger.info("Row skipped,file might be corrupted. Error {}".format(e))
@@ -62,15 +76,15 @@ def float_convert(val):
         pass
 
 
-def searchForParameters(elements, paramArray, exp_data=None):
-    '''
+def searchForParameters(elements, param_array, exp_data=None):
+    """
     Function purpose: return all attributes of a elements in the first row of a file \n
     Function purpose: return all attributes of a elements in the first row of a file \n
     Input: paramArray, elements \n
     Output: getAllRowElements(x,elements)
-    '''
+    """
     matching_elements = []
-    for x in paramArray:
+    for x in param_array:
         for row in elements[0]:
             p = re.compile(x, re.IGNORECASE)
             if p.search(row) is not None:
@@ -86,11 +100,11 @@ def searchForParameters(elements, paramArray, exp_data=None):
 
 
 def transformingIntoWGS84(crs, coordinate):
-    '''
+    """
     Function purpose: transforming SRS into WGS84 (EPSG:4326) \n
     Input: crs, point \n
     Output: retPoint constisting of x2, y2 (transformed points)
-    '''
+    """
     # TODO: check whether current src is 4326
     source = osr.SpatialReference()
     source.ImportFromEPSG(int(crs))
@@ -111,11 +125,11 @@ def transformingIntoWGS84(crs, coordinate):
 
 
 def transformingArrayIntoWGS84(crs, pointArray):
-    '''
-    Function purpose: transforming SRS into WGS84 (EPSG 4326; used by the GPS satellite navigation system) from an array \n
+    """
+    Function purpose: transforming SRS into WGS84 (EPSG 4326) from an array
     Input: crs, pointArray \n
     Output: array array
-    '''
+    """
     # print("----<>", pointArray)#
     array = []
     # vector_rep
@@ -132,8 +146,9 @@ def transformingArrayIntoWGS84(crs, pointArray):
 
 def validate_bbox_wgs84(bbox):
     """
-    :param bbox:
-    :return:
+    Function purpose: Validate if bbox is correct for WGS84
+    bbox: bounding box (list)
+    Output: True if bbox is correct for WGS84
     """
     valid = True
     lon_values = bbox[0:3:2]
@@ -148,8 +163,8 @@ def validate_bbox_wgs84(bbox):
 
 def flip_bbox(bbox):
     """
-    :param bbox:
-    :return:
+    bbox: Bounding box (list)
+    Output: bbox flipped (Latitude to longitude if possible)
     """
     # Flip values
     lon_values = bbox[1:4:2]
@@ -180,12 +195,12 @@ def getDelimiter(csv_file):
 
 
 def get_time_format(time_list, num_sample):
-    '''
+    """
     Function purpose: 'Guess' time format of a list of 'strings' by taking a representative sample
     time_list:  list of strings \n
     num_sample: size of the sample to determine time format \n
     Output: time format in string format (e.g '%Y.%M.d')
-    '''
+    """
 
     date_time_format = None
 
@@ -229,11 +244,11 @@ def get_time_format(time_list, num_sample):
 
 
 def date_parser(datetime_list, num_sample=None):
-    '''
+    """
     Function purpose: transform list of strings into date-time format
     datetime_list: list of date-times (strings) \n
     Output: list of DatetimeIndex
-    '''
+    """
 
     datetime_format = get_time_format(datetime_list, num_sample)
 
@@ -246,23 +261,29 @@ def date_parser(datetime_list, num_sample=None):
     return parse_time
 
 
-def extract_zip(zippedFile):
-    '''
+def extract_zip(filepath):
+    """
     Function purpose: unzip file (always inside a new folder)
-    Input: filepath
-    '''
+    filepath: filepath to zipfile
+    """
 
-    abs_path = os.path.abspath(zippedFile)
+    abs_path = os.path.abspath(filepath)
     root_folder = os.path.split(abs_path)[0]
     zip_name = os.path.split(abs_path)[1][:-4]
     zip_folder_path = os.path.join(root_folder, zip_name)
 
-    with zipfile.ZipFile(abs_path) as zipf:
-        zipf.extractall(zip_folder_path)
+    with zipfile.ZipFile(abs_path) as zip_file:
+        zip_file.extractall(zip_folder_path)
 
 
 def bbox_merge(metadata, origin):
-    logger.debug("medatada {}".format(metadata))
+    """
+    Function purpose: merge bounding boxes
+    metadata: metadata with geoextent extraction from multiple files (dict)
+    origin: folder path or filepath (str)
+    Output: Merged bbox (dict)
+    """
+    logger.debug("metadata {}".format(metadata))
     boxes_extent = []
     metadata_merge = {}
     num_files = len(metadata.items())
@@ -307,7 +328,7 @@ def bbox_merge(metadata, origin):
 
             except Exception as e:
                 logger.debug(
-                    "Error extracting geographic extent of {}. CRS {} may be invalid. Error: {}".format(x, bbox[1], e))
+                    "Error extracting geographic extent. CRS {} may be invalid. Error: {}".format(int(bbox[1]), e))
                 continue
 
         num_geo_files = multipolygon.GetGeometryCount() / 4
@@ -325,6 +346,12 @@ def bbox_merge(metadata, origin):
 
 
 def tbox_merge(metadata, path):
+    """
+    Function purpose: Merge time boxes
+    metadata: metadata with geoextent extraction from multiple files (dict)
+    path: path of directory being merged
+    Output: Merged tbox
+    """
     boxes = []
     num_files = len(metadata.items())
     for x, y in metadata.items():
@@ -350,3 +377,202 @@ def tbox_merge(metadata, path):
         time_ext = [min_date, max_date]
 
     return time_ext
+
+
+def transform_bbox(x):
+    """
+    Function purpose: Transform bounding box (str) into geometry
+    x: bounding box (str)
+    """
+
+    try:
+        ring = ogr.Geometry(ogr.wkbLinearRing)
+        ring.AddPoint(x[0], x[1])
+        ring.AddPoint(x[2], x[1])
+        ring.AddPoint(x[0], x[3])
+        ring.AddPoint(x[2], x[3])
+        ring.CloseRings()
+    # Create polygon
+        poly = ogr.Geometry(ogr.wkbPolygon)
+        poly.AddGeometry(ring)
+        poly.FlattenTo2D()
+        bbox = poly.ExportToWkt()
+
+    except:
+
+        bbox = None
+
+    return bbox
+
+
+def transform_tbox(x):
+    """
+    Function purpose: Transform time box (list) into int
+    x: time box (list)
+    """
+
+    if x is None:
+        return None
+    elif isinstance(x, list):
+        return str(x[0]) + '/' + str(x[1])
+
+
+def extract_details(details):
+    """
+    Function purpose: Extracts details from geoextent extraction
+    details: dictionary with geoextent extraction
+    Output: dataframe organized by filename, file format, handler, bbox, tbox and crs by file.
+    """
+
+    filename = []
+    file_format = []
+    handler = []
+    bbox = []
+    tbox = []
+    crs = []
+
+    for i in details:
+
+        file = details[i]
+
+        if file is None:
+            filename.append([i])
+            file_format_v = os.path.splitext(i)[1][1:]
+            if file_format_v == '':
+                file_format_v = 'undetected'
+            file_format.append([file_format_v])
+            handler.append([None])
+            bbox.append([None])
+            tbox.append([None])
+            crs.append([None])
+        else:
+            filename.append([i])
+            file_format.append([file.get('format')])
+            handler_v = file.get('geoextent_handler')
+            bbox_v = file.get('bbox')
+            tbox_v = file.get('tbox')
+            crs_v = file.get('crs')
+            handler.append([handler_v])
+            bbox.append([bbox_v])
+            tbox.append([tbox_v])
+            crs.append([crs_v])
+
+            if file.get('format') == 'folder':
+                details_folder = extract_details(file['details'])
+                filename.append(details_folder['filename'])
+                file_format.append(details_folder['format'])
+                handler.append(details_folder['handler'])
+                bbox.append(details_folder['bbox'])
+                tbox.append(details_folder['tbox'])
+                crs.append(details_folder['crs'])
+
+    if any(isinstance(i, list) for i in filename):
+        filename = list(itertools.chain.from_iterable(filename))
+        file_format = list(itertools.chain.from_iterable(file_format))
+        handler = list(itertools.chain.from_iterable(handler))
+        bbox = list(itertools.chain.from_iterable(bbox))
+        tbox = list(itertools.chain.from_iterable(tbox))
+        crs = list(itertools.chain.from_iterable(crs))
+
+    d = {'filename': filename, 'format': file_format,   'handler': handler,
+         'bbox': bbox,
+         'tbox': tbox, 'crs': crs}
+    files = pd.DataFrame(d)
+    return files
+
+
+def extract_output(result, files, current_version):
+    """
+    Function purpose: Extracts final output from geoextent including all files and containing folder
+    result: geoextent output from extraction
+    files: user input for initial extraction (e.g name of the main folder)
+    current_version: Current geoextent version
+    Output: Dataframe with geoextent of all files AND final output (merge) of user request
+    """
+    filename = files
+    file_format = result.get('format')
+    handler = "geoextent:" + current_version
+    bbox = result.get('bbox')
+    tbox = result.get('tbox')
+    crs = result.get('crs')
+
+    new_row = {'filename': filename, 'format': file_format,   'handler': handler, 'bbox': bbox, 'tbox': tbox, 'crs': crs
+               }
+
+    df = extract_details(result['details'])
+    df = df.append(new_row, ignore_index=True)
+    df['bbox'] = df['bbox'].apply(transform_bbox)
+    df['tbox'] = df['tbox'].apply(transform_tbox)
+    return df
+
+
+def is_doi(val):
+    """
+    Function purpose: Returns None if val doesn't match pattern of a DOI.
+    http://en.wikipedia.org/wiki/Digital_object_identifier.
+    """
+    return doi_regexp.match(val)
+
+
+def normalize_doi(val):
+    """
+    Function purpose: Return just the DOI (e.g. 10.1234/jshd123)
+    from a val that could include a url or doi
+    (e.g. https://doi.org/10.1234/jshd123)
+    val: DOI or URL (str)
+    """
+    m = doi_regexp.match(val)
+    return m.group(2)
+
+
+def create_geopackage(df, filename):
+    """
+    Function purpose: Creates a geopackage file
+    df: dataframe from extract_output result
+    filename: Name for the Geopackage file
+    """
+    sr4326 = osr.SpatialReference()
+    sr4326.ImportFromEPSG(WGS84_EPSG_ID)
+
+    if os.path.exists(filename):
+        os.remove(filename)
+        logger.warning("Overwriting {} ".format(filename))
+
+    ds = ogr.GetDriverByName('GPKG').CreateDataSource(filename)
+    lyr = ds.CreateLayer('files', geom_type=ogr.wkbPolygon, srs=sr4326)
+    lyr.CreateField(ogr.FieldDefn('filename', ogr.OFTString))
+    lyr.CreateField(ogr.FieldDefn('handler', ogr.OFTString))
+    lyr.CreateField(ogr.FieldDefn('format', ogr.OFTString))
+    lyr.CreateField(ogr.FieldDefn('tbox', ogr.OFTString))
+    lyr.CreateField(ogr.FieldDefn('crs', ogr.OFTString))
+
+    for i in range(len(df)):
+        feat = ogr.Feature(lyr.GetLayerDefn())
+        feat['filename'] = df.loc[i, "filename"]
+        feat['format'] = df.loc[i, "format"]
+        feat['tbox'] = df.loc[i, "tbox"]
+        feat['handler'] = df.loc[i, "handler"]
+        feat['crs'] = df.loc[i, "crs"]
+        if df.loc[i, "bbox"] is not None:
+            feat.SetGeometry(ogr.CreateGeometryFromWkt(df.loc[i, "bbox"]))
+        lyr.CreateFeature(feat)
+
+    ds = None
+
+
+def path_output(path):
+
+    if os.path.isdir(path):
+        logger.error("Output must be a file, not a directory ")
+        raise ValueError("Output must be a file, not a directory: {}".format(path))
+
+    folder_path = os.path.split(path)[0]
+    user_path = Path(folder_path)
+    if user_path.exists():
+        absolute_file_path = user_path.as_posix() + "/" + os.path.split(path)[1]
+    else:
+        logger.error("Output target directory does not exist: {}".format(path))
+        raise ValueError("Output target directory does not exist: {}".format(path))
+    return absolute_file_path
+
+